Fix issue of closed file descriptor
Calling fclose on a FILE* closes the underlying file descriptor which
was associated with the FILE* by calling fdopen. This is undesirable in
case the file descriptor is meant to be reused. Duplicate the original
file descriptor using dup before calling fdopen. User retains the
responsibilty to manage the file descriptor.